+++ src/dj_hyperview/__init__.py
@@ -43,7 +43,7 @@ from .views import HyperviewTemplateView
 try:
     __version__ = _distribution_version("dj-hyperview")
 except _PackageNotFoundError:
-    __version__ = "0.1.0a13"
+    __version__ = "0.1.0a14"
 
 __all__ = [
     "__version__",